Implementation of dwalk
dwalk is the module in charge of parallel walking of the filesystem.

I looked at other go modules for parallel filesystem walking but decided
to implement my own for the time being. Maybe in the future I will end
up using gowalker or saracen/walker.

Currently, dwalk is somewhat minimal. Benchmarks need to be done and
obviously some major refactoring.