DMap implemented. Began to implement monitor.
Dmap structure was created. This is a wrapper structure over
a hash map. Dmap methods:

- Add()
- MapSize()
- FileCount()
- Get()

=========

Began putting together the monitor loop in main.go. This is the core
loop that reads from the worker go routines via dFile channel and adds
the content to the Dmap for duplication detection.

main.go monitor loop TODO:

- Finish dMap.Add() call.
- Add default select case.
- Test cancellation.