Modified Dfile structure and methods.
Dfile constructor NewDfile() now takes file size instead of calculating
it. This is because our call to ReadDir will supply the files sizes in
the os.FileInfo array it returns. This saves us some work.

Added Dfile.BaseName() method. This returns the base name of the file
while Dfile.FileName() returns the fully qualified name of the file
(the file name with its path).