Added result tree using Pterm.
Displayed results are now shown as a pretty tree. Modifications
still need to be made for making results easier to read.

Added TODO file. The file contains items that need to be completed (in
nor particular order as of now). Reference this file to see what tasks
need to be worked on.