Added command line flags. Total elapsed time.
Added two flags. One to suppress output of banner. The other
is time in miliseconds that is used to update user about progress.

I also added code to measure total time the scan took. It is displayed
to the user at the end when it prints the results.

Updated TODO.txt