Fixed open file descriptor issue. Changed logging.
Previously running dskDitto on a directory with tons of files
exhausted the max file descriptor limit which is only 256 by default
on macOS. In the future we may want to raise this limit. However,
a semaphore was added to the MD5 hashing function to ensure we
don't max out on open files. The current value is set to 100
but it we will add ability to set limit from command line options.

Since the logging to stderr/stdout was giving too much verbose info, we
now log to files. A bit more appropriate anyhow.

Added a PerformanceLog.md that is a simple journal of dskDitto runs,
the time it took and all that stuff. It's just meant to serve as a rough
guide to track performance.

Next issue we need to address is the atroicious performance of Pterm
when it dumps the final results. It took more time to display the
results as a tree than it did to scan almost 1 million files
(see PerformanceLog.md Run #1)!